appmanager cleanup
