+++ server/v2/appmanager/mock/storage.go
@@ -18,7 +18,7 @@ func StateCommitment(_ *testing.T) storev2.Committer {
 	db := db.NewMemDB()
 	tree := iavl.NewIavlTree(db, logger{}, iavl.DefaultConfig())
 
-	sc, _ := commitment.NewCommitStore(map[string]commitment.Tree{"": tree}, db, logger{})
+	sc, _ := commitment.NewCommitStore(map[string]commitment.Tree{"": tree}, db, nil, logger{})
 	return sc
 }
 
@@ -26,7 +26,7 @@ func StateStorage(t *testing.T) storev2.VersionedDatabase {
 	t.Helper()
 	db, err := pebbledb.New(t.TempDir())
 	require.NoError(t, err)
-	return storage.NewStorageStore(db)
+	return storage.NewStorageStore(db, nil, logger{})
 }
 
 type logger struct{}
