+++ server/v2/core/go.mod
@@ -7,6 +7,7 @@ replace cosmossdk.io/core => ../../../core
 require (
 	cosmossdk.io/core v0.11.0
 	github.com/cosmos/gogoproto v1.4.11
+	github.com/cosmos/ics23/go v0.10.0
 	google.golang.org/protobuf v1.32.0
 )
 
@@ -18,6 +19,7 @@ require (
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/rs/zerolog v1.32.0 // indirect
+	golang.org/x/crypto v0.19.0 // indirect
 	golang.org/x/exp v0.0.0-20231006140011-7918f672742d // indirect
 	golang.org/x/net v0.21.0 // indirect
 	golang.org/x/sys v0.17.0 // indirect
